requirements.txt
django==5.0.4
psycopg2-binary==2.9.9
djangorestframework==3.15.1
python-dotenv==1.0.1
cryptography==42.0.5